requirements.txt
# Alpaca API
alpaca-py>=0.21.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Telegram notifications
python-telegram-bot>=20.0

# Scheduler
APScheduler>=3.10.0

# Logging
loguru>=0.7.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Timezone handling
pytz>=2024.1
